requirements.txt
openai==0.27.4
gradio==3.24.1
python-dotenv